backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
passlib
pandas
openpyxl
bcrypt
